Temporary Fix For AOT Compiler Issue
This is a pretty dirty hack for the dart AOT issue causing `type '_OneByteString' is not a subtype of type 'int' of 'index’`. I am basically parsing out the identity id from the response string instead of getting it from the map until we figure out what the actual issue is.

* refs: furaiev/amazon-cognito-identity-dart-2#43
* refs: furaiev/amazon-cognito-identity-dart-2#26